updated top level scripts to allow substitution of variables
updated all scripts to assume defaults for dependent variables, to avoid
failure when executed stand-alone

+++ scripts/linux/comm/weechat.sh
@@ -3,9 +3,14 @@
 # install package
 aptitude install -ryq weechat
 
-# generate weechat config with first-run
-su $username -s /bin/bash -c 'weechat-curses &>/dev/null & pid=$!; sleep 10; kill -9 $pid'
+# try to setup some values for our user
+if [ -n "$username" ]
+then
 
-# configure weechat
-sed -i 's/max_buffer_lines_number.*/max_buffer_lines_number = 0/' "/home/${username}/.weechat/weechat.conf"
-sed -i 's/freenode\.autoconnect.*/freenode\.autoconnect = on/' "/home/${username}/.weechat/irc.conf"
+    # generate weechat config with first-run
+    su $username -s /bin/bash -c 'weechat-curses &>/dev/null & pid=$!; sleep 10; kill -9 $pid'
+
+    # configure weechat
+    sed -i 's/max_buffer_lines_number.*/max_buffer_lines_number = 0/' "/home/${username}/.weechat/weechat.conf"
+    sed -i 's/freenode\.autoconnect.*/freenode\.autoconnect = on/' "/home/${username}/.weechat/irc.conf"
+fi
